add matrix bloc extraction and convolution product add edge detection and bluring frontends.

+++ src/tests/mult3x3.c
@@ -5,14 +5,14 @@
 int main() {
   uint16_t n = 3;
   uint16_t m = 3;
-  uint16_t weights[9] = {
+  int32_t weights[9] = {
                           1, 2, 3,
                           3, 1, 2,
                           2, 3, 1
                         };
   struct matrix *a = matrix_from(n, m, weights);
 
-  uint16_t weights2[9] = {
+  int32_t weights2[9] = {
                           1, 2, 4,
                           8, 16, 32,
                           64, 128, 255,
@@ -22,7 +22,7 @@ int main() {
   struct matrix *c = matrix_new(n, m);
   matmult3x3(a, b, c);
 
-  uint16_t exp_weights[9] = {
+  int32_t exp_weights[9] = {
                             209, 418, 833,
                             139, 278, 554,
                             90, 180, 359
